% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Present value function}
\usage{
pv(FV, r, n = 5)
}
\arguments{
\item{FV}{a numeric vector of future values}

\item{r}{the discount rate}

\item{n}{= number of time steps between present and future}
}
\value{
an numeric vector of present values
}
\description{
conomputes the present value of a sum given a future sum
}
