% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_my_app}
\alias{run_my_app}
\title{Runs first demo app in package}
\usage{
run_my_app(app_name, ...)
}
\arguments{
\item{app_name}{character string for a directory in this package}

\item{...}{Addt'l options passed to shinyAppDir}
}
\value{
A printed shiny app
}
\description{
Goofin' around, learnin' to make packages and apps
}
\examples{
\dontrun{ run_my_app('myFirstApp')}

}
